Updated info about bulbs/group
I could get my bulbs detected as per the setup guide, however they didnt work.
They showed up in the UI, but would not react to the triggers. 

Tried removing and uninstalling, tried rebooting everything, changing if defined or not in configuration.

What made it work for me was the configuration of names of the bulbs and group.. 
So I think maybe the standard names were either too long or that it got confused somehow.